Split off the upgrade views from the application instance admin pages
